+++ Fun.Blazor/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Fix hidden attribute
+
 ## [3.2.0] - 2023-11-16
 
 - Unify version